travis updates

+++ cbh_tests/features/environment.py
@@ -251,22 +251,29 @@ class Tester(unittest.TestCase):
 import os
 CONDA_ENV_PATH = os.getenv("CONDA_ENV_PATH")
 host = CONDA_ENV_PATH + "/var/postgressocket"
+database_path = CONDA_ENV_PATH + "/var/"
 
 def before_all(context):
     from cbh_datastore_ws.features.steps.datastore_realdata import create_realdata, project
     from subprocess import Popen, PIPE, call
-    call(
-        "dropdb dev_db --if-exists -h %s" % host , shell=True)
-    call(
-        "createdb dev_db -h %s" % host, shell=True)
+    process = Popen(['git', 'log', '--format="%H"', '-n', '1'], stdout=PIPE)
+    commit_based_filename, error = process.communicate()
+    commit_based_filename = database_path + str(commit_based_filename[1:-2])
+    print(commit_based_filename)
 
-    call('psql -h %s -c "create extension hstore;create extension rdkit;" dev_db' % host, shell=True)
+    if not os.path.isfile(commit_based_filename):
+        call(
+            "dropdb dev_db --if-exists -h %s" % host , shell=True)
+        call(
+            "createdb dev_db -h %s" % host, shell=True)
 
-    call(
-        "python manage.py migrate", shell=True)   
-    call(
-        "pg_dump dev_db -Fc -h %s > /tmp/mydevdb.dump" % host , shell=True)    
-    
+        call('psql -h %s -c "create extension hstore;create extension rdkit;" dev_db' % host, shell=True)
+
+        call(
+            "python manage.py migrate", shell=True)   
+        call(
+            "pg_dump dev_db -Fc -h %s > %s" % (host, commit_based_filename) , shell=True)    
+        
     pass
     # Even though DJANGO_SETTINGS_MODULE is set, this may still be
     # necessary. Or it may be simple CYA insurance.
@@ -291,6 +298,7 @@ def before_all(context):
 def before_scenario(context, scenario):
     # Set up the scenario test environment
     from subprocess import Popen, PIPE, call
+
 
     call(
         "dropdb dev_db --if-exists -h %s" % host , shell=True)
